# docs/swagger/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: 项目管理平台 Ydsz Plane 的 REST API（需求/任务/缺陷/迭代/版本/自动化/Webhook/知识库/仪表盘/全局搜索/通知等）。
  title: Ydsz Plane API
  version: 1.0.0
paths:
...
  /knowledge/search:
    get:
      description: 跨知识库空间全文检索（PostgreSQL tsvector）
      parameters:
      - description: 关键词
        in: query
        name: q
        required: true
        type: string
      - description: 空间 ID（可选）
        in: query
        name: space_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 知识库检索
      tags:
      - knowledge
...
  /search:
    get:
      description: |-
        跨对象（issues/sprints/versions）全文检索 + 高亮 + JQL 结构化过滤
        JQL 示例: project:YD type:defect priority:high assignee:me()
      parameters:
      - description: 搜索词（支持类 JQL 语法）
        in: query
        name: q
        required: true
        type: string
      - description: 过滤类型 (issue|sprint|version)
        in: query
        name: doc_type
        type: string
      - description: 工作项类型 (epic|requirement|task|defect)
        in: query
        name: type
        type: string
      - description: 优先级
        in: query
        name: priority
        type: string
      - description: 状态 ID
        in: query
        name: state_id
        type: integer
      - description: 每页数量 (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 全文搜索（支持类 JQL 语法）
      tags:
      - search
...
securityDefinitions:
  Bearer:
    description: '"格式: Bearer {access_token}。登录/注册接口返回 access_token 后填入。"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"